CAMEL-19279: make token directly nested conf element
