+++ components/camel-azure/camel-azure-files/src/main/java/org/apache/camel/component/file/azure/FilesConfiguration.java
@@ -28,36 +28,8 @@ public class FilesConfiguration extends RemoteFileConfiguration {
     public static final int DEFAULT_HTTPS_PORT = 443;
     public static final String DEFAULT_INTERNET_DOMAIN = "file.core.windows.net";
 
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String sv;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String ss;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String srt;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String sp;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String se;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String st;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String spr;
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    protected String sig;
-
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    private String si;  // service SAS only
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    private String sr;  // service SAS only
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    private String sdd; // service SAS only
-    @UriParam(label = "both", description = "part of SAS token", secret = true)
-    private String sip;
-
     @UriParam(label = "both", description = "Shared key (storage account key)", secret = true)
     private String sharedKey;
-
-    private FilesToken token = new FilesToken();
 
     private String account;
     private String share;
@@ -117,124 +89,12 @@ public class FilesConfiguration extends RemoteFileConfiguration {
         return account;
     }
 
-    public String getSv() {
-        return sv;
-    }
-
-    public void setSv(String sv) {
-        token.setSv(sv);
-        this.sv = sv;
-    }
-
-    public String getSs() {
-        return ss;
-    }
-
-    public void setSs(String ss) {
-        token.setSs(ss);
-        this.ss = ss;
-    }
-
-    public String getSrt() {
-        return srt;
-    }
-
-    public void setSrt(String srt) {
-        token.setSrt(srt);
-        this.srt = srt;
-    }
-
-    public String getSp() {
-        return sp;
-    }
-
-    public void setSp(String sp) {
-        token.setSp(sp);
-        this.sp = sp;
-    }
-
-    public String getSe() {
-        return se;
-    }
-
-    public void setSe(String se) {
-        token.setSe(se);
-        this.se = se;
-    }
-
-    public String getSt() {
-        return st;
-    }
-
-    public void setSt(String st) {
-        token.setSt(st);
-        this.st = st;
-    }
-
-    public String getSpr() {
-        return spr;
-    }
-
-    public void setSpr(String spr) {
-        token.setSpr(spr);
-        this.spr = spr;
-    }
-
-    public String getSig() {
-        return sig;
-    }
-
-    public void setSig(String sig) {
-        token.setSig(sig);
-        this.sig = sig;
-    }
-
-    public String getSi() {
-        return si;
-    }
-
-    public void setSi(String si) {
-        token.setSi(si);
-        this.si = si;
-    }
-
-    public String getSr() {
-        return sr;
-    }
-
-    public void setSr(String sr) {
-        token.setSr(sr);
-        this.sr = sr;
-    }
-
-    public String getSdd() {
-        return sdd;
-    }
-
-    public void setSdd(String sdd) {
-        token.setSdd(sdd);
-        this.sdd = sdd;
-    }
-
-    public String getSip() {
-        return sip;
-    }
-
-    public void setSip(String sip) {
-        token.setSip(sip);
-        this.sip = sip;
-    }
-
     public String getSharedKey() {
         return sharedKey;
     }
 
     public void setSharedKey(String sharedKey) {
         this.sharedKey = sharedKey;
-    }
-
-    public FilesToken getToken() {
-        return token;
     }
 
 }
